requirements.txt
Flask==2.2.5
flask-mongoengine==1.0.0
mongoengine==0.27.0
pymongo<4.0.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SocketIO==5.3.4
cryptography==41.0.3
redis==5.0.0
flask-caching==2.0.2
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
greenlet==3.0.3
gevent==23.9.1
gevent-websocket==0.10.1
dnspython<2.3.0 
Werkzeug==2.2.3
bcrypt==4.0.1
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
pytz==2024.1
APScheduler==3.10.4
docker==7.1.0


